% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_bibtex_string.R
\name{append_bibtex_entry}
\alias{append_bibtex_entry}
\title{Check whether a tag from a bibtex exists in a set of existing content.}
\usage{
append_bibtex_entry(bibtex_entry, existing_content, verbose = FALSE)
}
\arguments{
\item{bibtex_entry}{a string of bibtex}

\item{existing_content}{a vector of strings to check against}

\item{verbose}{whether to print a message if the tag already exists}
}
\value{
the existing content with the bibtex string appended if it does not already exist
}
\description{
Check whether a tag from a bibtex exists in a set of existing content.
}
\examples{

append_bibtex_entry(
bibtex_entry = v_example_refs[1],
existing_content = v_example_refs[1],
verbose = TRUE
)

append_bibtex_entry(
bibtex_entry = v_example_refs[1],
existing_content = v_example_refs[1],
verbose = FALSE
)
}
