% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_bibtex_string.R
\name{append_bibtex_entry_vec}
\alias{append_bibtex_entry_vec}
\title{Append the bibtex file for a series (vector of bibtex strings)}
\usage{
append_bibtex_entry_vec(v_bibtex_entries, existing_content, verbose = FALSE)
}
\arguments{
\item{v_bibtex_entries}{a vector of bibtex strings}

\item{existing_content}{the existing content, generated by ReadLines in R on the .bib file.}

\item{verbose}{whether to print messages to the console}
}
\value{
a vector of bibtex content
}
\description{
If the tag already exists in the content then the bibtex string is ignored
otherwise it is incorporated into the content at the bottom, with a space between
}
\examples{

append_bibtex_entry_vec(
v_bibtex_entries = v_example_refs[2],
existing_content = v_example_refs[1:5],
verbose = TRUE
)

append_bibtex_entry_vec(
v_bibtex_entries = v_example_refs[1],
existing_content = v_example_refs[1:5],
verbose = FALSE
)
}
