% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_bibtex_string.R
\name{append_bibtex_string}
\alias{append_bibtex_string}
\title{Check whether a tag from a bibtex exists in a set of existing content.}
\usage{
append_bibtex_string(bibtex_string, existing_content)
}
\arguments{
\item{bibtex_string}{a string of bibtex}

\item{existing_content}{a vector of strings to check against}
}
\value{
the existing content with the bibtex string appended if it does not already exist
}
\description{
Check whether a tag from a bibtex exists in a set of existing content.
}
\examples{

append_bibtex_string(bibtex_string = v_example_refs[2], existing_content = v_example_refs[1])
}
