% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_bibtex_string.R
\name{append_bibtex_string_vec}
\alias{append_bibtex_string_vec}
\title{Append the bibtex file for a series (vector of bibtex strings)}
\usage{
append_bibtex_string_vec(v_bibtex_strings, existing_content)
}
\arguments{
\item{v_bibtex_strings}{a vector of bibtex strings}

\item{existing_content}{the existing content, generated by ReadLines in R on the .bib file.}
}
\value{
a vector of bibtex content
}
\description{
If the tag already exists in the content then the bibtex string is ignored
otherwise it is incorporated into the content at the bottom, with a space between
}
