% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tag_format.R
\name{convert_tag_format}
\alias{convert_tag_format}
\title{Convert between different tag formats}
\usage{
convert_tag_format(x, to = c("markdown_long", "markdown_short", "tag"))
}
\arguments{
\item{x}{a character vector of tags}

\item{to}{the format to convert to}
}
\value{
a character vector of the tags in the new format
}
\description{
Convert between different tag formats
}
\examples{
convert_tag_format(x = "@smith2020making", to = "tag")

}
