% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bib_if_missing.R
\name{create_bib_if_missing}
\alias{create_bib_if_missing}
\title{Create empty bib file if it is missing}
\usage{
create_bib_if_missing(path)
}
\arguments{
\item{path}{path to file. Must end with .bib}
}
\value{
empty
}
\description{
Create empty bib file if it is missing
}
\examples{
\dontrun{
create_bib_if_missing("refs.bib")
}
}
