% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bibtex_tag.R
\name{extract_bibtex_tag}
\alias{extract_bibtex_tag}
\title{Extracts the tag from the bibtex reference}
\usage{
extract_bibtex_tag(
  bibtex_string,
  output = c("tag", "markdown_short", "markdown_long")
)
}
\arguments{
\item{bibtex_string}{the bibtex string to use}

\item{output}{type of tag to output (e.g. markdown short ref or long reference or just the tag string)}
}
\value{
a bibtex tag, or pre-formatted for markdown
}
\description{
Exists after the first braket and before the first comma.
}
\examples{

extract_bibtex_tag("@article{smith2020making,
title={Making health economic models Shiny: A tutorial},
author={Smith, Robert and Schneider, Paul},
journal={Wellcome open research},
volume={5},
pages={69},
year={2020}
}", "markdown_long"
  )

extract_bibtex_tag("Assumption")
}
