% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bib_file_from_vector.R
\name{update_bib_file_from_vector}
\alias{update_bib_file_from_vector}
\title{Update or create a bib file with a vector of bibtex strings}
\usage{
update_bib_file_from_vector(
  path_to_bib,
  v_bibtex_strings,
  output_type = "markdown_short"
)
}
\arguments{
\item{path_to_bib}{file path to the bib to create or update}

\item{v_bibtex_strings}{bibtex strings to add}

\item{output_type}{the type of output to return. Default is "markdown_short"}
}
\value{
a vector of the bibtex tags (one per string)
}
\description{
Update or create a bib file with a vector of bibtex strings
}
