% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_example_refs}
\alias{v_example_refs}
\title{Example bibtex dataset}
\format{
\subsection{\code{v_example_refs}}{

A vector containing 8 unique bibtex strings
}
}
\source{
None
}
\usage{
v_example_refs
}
\description{
A vector of bibtex strings
}
\keyword{datasets}
