% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bib_if_missing.R
\name{create_bib_if_missing}
\alias{create_bib_if_missing}
\title{Create empty bib file if it is missing}
\usage{
create_bib_if_missing(path = "bibby_refs.bib", verbose = FALSE)
}
\arguments{
\item{path}{path to proposed new file name. Must end with .bib}

\item{verbose}{logical for whether to print messages. Default is FALSE}
}
\value{
empty
}
\description{
Create empty bib file if it is missing
}
\examples{
\donttest{
create_bib_if_missing()
}
}
